+++ src/idlib/math/Vector.h
@@ -1123,7 +1123,7 @@ ID_INLINE idVec3 idVec3::Random( const idVec3& range, idRandom& random ) const {
 // RAVEN END
 
 ID_INLINE bool idVec3::IsZero( void ) const {
-	return ( ( ( *( const unsigned long * ) &( x ) ) | ( *( const unsigned long * ) &( y ) ) | ( *( const unsigned long * ) &( z ) ) ) & ~( 1<<31 ) ) == 0;
+	return ( ( ( *( const unsigned int * ) &( x ) ) | ( *( const unsigned int * ) &( y ) ) | ( *( const unsigned int * ) &( z ) ) ) & ~( 1<<31 ) ) == 0;
 }
 
 //===============================================================
